update 31-05
change Weight type from int to double as default

+++ src/main/java/com/example/SS2_Backend/dto/request/IndividualDeserializer.java
@@ -40,7 +40,7 @@ public class IndividualDeserializer extends StdDeserializer<Individual> {
 				JsonNode propertyNode = propertiesIterator.next();
 
 				Double propertyValue = propertyNode.get(2).asDouble();
-				int propertyWeight = propertyNode.get(1).asInt();
+				Double propertyWeight = propertyNode.get(1).asDouble();
 				String inputRequirement = propertyNode.get(0).asText();
 
 				individual.setProperty(propertyValue, propertyWeight, inputRequirement);
